+++ src/common/components/layout/PageSection.tsx
@@ -1,13 +1,13 @@
 import React, { ReactNode } from 'react';
 
-import styles from './youthProfileFormSection.module.css';
+import styles from './pageSection.module.css';
 
 interface Props {
   children: ReactNode;
 }
 
 function YouthProfileFormSection(props: Props) {
-  return <div className={styles.youthProfileFormSection} {...props} />;
+  return <div className={styles.pageSection} {...props} />;
 }
 
 export default YouthProfileFormSection;