+++ CHANGELOG.md
@@ -10,5 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - Controls for additional contact persons
+- List of additional contact persons in membership details view
+
+### Fixed
+- Authentication info text position
 
 ## [1.0.0-rc.3]